# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "znskr api",
    "description": "znskr paas api for managing apps and deployments",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/apps/{app_id}/deployments/{id}": {
      "get": {
        "tags": [
          "deployments"
        ],
        "summary": "get a single deployment",
        "operationId": "get_deployment",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "deployment id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deployment details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/apps/{app_id}/deployments/{id}/logs": {
      "get": {
        "tags": [
          "deployments"
        ],
        "summary": "get deployment logs",
        "operationId": "get_deployment_logs",
        "parameters": [
          {
            "name": "app_id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "deployment id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "lines limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "lines offset",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deployment logs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/github/app": {
      "get": {
        "tags": [
          "github-app"
        ],
        "summary": "get github app status and installations",
        "operationId": "get_github_app",
        "responses": {
          "200": {
            "description": "github app status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubAppStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "github-app"
        ],
        "summary": "delete github app configuration",
        "operationId": "delete_github_app",
        "responses": {
          "200": {
            "description": "github app deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AppDetails": {
        "type": "object",
        "description": "app details",
        "required": [
          "app_id",
          "app_name",
          "html_url"
        ],
        "properties": {
          "app_id": {
            "type": "integer",
            "format": "int64"
          },
          "app_name": {
            "type": "string"
          },
          "html_url": {
            "type": "string"
          }
        }
      },
...
      "DeploymentResponse": {
        "type": "object",
        "description": "deployment response",
        "required": [
          "id",
          "app_id",
          "commit_sha",
          "status",
          "created_at"
        ],
        "properties": {
          "app_id": {
            "type": "string",
            "format": "uuid"
          },
          "commit_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "commit_sha": {
            "type": "string"
          },
          "container_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "finished_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "error message"
          }
        }
      },
...
      "GithubAppStatusResponse": {
        "type": "object",
        "description": "github app status response",
        "required": [
          "configured",
          "installations"
        ],
        "properties": {
          "app": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AppDetails",
                "description": "app details if configured"
              }
            ]
          },
          "configured": {
            "type": "boolean",
            "description": "whether a github app is configured"
          },
          "installations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstallationDetails"
            },
            "description": "list of installations"
          }
        }
      },
...
      "InstallationDetails": {
        "type": "object",
        "description": "installation details",
        "required": [
          "id",
          "account_login",
          "account_type"
        ],
        "properties": {
          "account_login": {
            "type": "string"
          },
          "account_type": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "repository_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}